Add calc, eval, greeter, PCC roles; work on info, voiceText

+++ src/main/java/modulebot/voiceText/Main.java
@@ -5,27 +5,26 @@ import modulebot.main.hosts.CommandHost;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.*;
 import net.dv8tion.jda.core.events.ReadyEvent;
-import net.dv8tion.jda.core.events.guild.member.GuildMemberJoinEvent;
-import net.dv8tion.jda.core.events.guild.member.GuildMemberLeaveEvent;
 import net.dv8tion.jda.core.events.guild.voice.GuildVoiceJoinEvent;
 import net.dv8tion.jda.core.events.guild.voice.GuildVoiceLeaveEvent;
 import net.dv8tion.jda.core.events.guild.voice.GuildVoiceMoveEvent;
-import org.json.JSONObject;
+import net.dv8tion.jda.core.requests.restaction.ChannelAction;
 
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
 
 public class Main extends CommandHost {
-    // TODO: remove v
-    private static JSONObject roles;
-    private static HashMap<Long, HashMap<String, ArrayList<String>>> settings = new HashMap<>();
-    // TODO: remove ^
+    static HashMap<Long, HashMap<String, ArrayList<String>>> settings = new HashMap<>();
 
     @Override
     public Command[] getCommands() {
-        return new Command[0];
+        return new Command[] { new Configure() };
     }
 
     @Override
@@ -40,29 +39,28 @@ public class Main extends CommandHost {
 
     @Override
     public void onReady(ReadyEvent event) {
-        System.out.println("voiceText READY EVENT");
         try {
             Statement st = modulebot.main.Main.conn.createStatement();
             st.execute("" +
                     "create table if not exists voiceText(" +
-                        "id       bigint unsigned not null," +
-                        "banUsers text            null," +
-                        "banRoles text            null," +
-                        "admUsers text            null," +
-                        "admRoles text            null" +
+                        "id            bigint unsigned not null unique," +
+                        "disallowUsers text            null," +
+                        "disallowRoles text            null," +
+                        "alwaysUsers   text            null," +
+                        "alwaysRoles   text            null" +
                     ") COLLATE utf8_bin engine InnoDB");
             ResultSet rs = st.executeQuery("SELECT * FROM voiceText");
             while (rs.next()) {
                 long id = rs.getLong("id");
                 HashMap<String, ArrayList<String>> map = new HashMap<>();
-                String banUsers = rs.getString("banUsers");
-                String banRoles = rs.getString("banRoles");
-                String admUsers = rs.getString("admUsers");
-                String admRoles = rs.getString("admRoles");
-                map.put("banUsers", new ArrayList<>(Arrays.asList((banUsers == null ? ""     : banUsers).split(";"))));
-                map.put("banRoles", new ArrayList<>(Arrays.asList((banRoles == null ? ""     : banRoles).split(";"))));
-                map.put("admUsers", new ArrayList<>(Arrays.asList((admUsers == null ? ""     : admUsers).split(";"))));
-                map.put("admRoles", new ArrayList<>(Arrays.asList((admRoles == null ? ""     : admRoles).split(";"))));
+                String disallowUsers = rs.getString("disallowUsers");
+                String disallowRoles = rs.getString("disallowRoles");
+                String alwaysUsers   = rs.getString("alwaysUsers");
+                String alwaysRoles   = rs.getString("alwaysRoles");
+                map.put("disallowUsers", new ArrayList<>(Arrays.asList((disallowUsers == null ? "" : disallowUsers).split(";"))));
+                map.put("disallowRoles", new ArrayList<>(Arrays.asList((disallowRoles == null ? "" : disallowRoles).split(";"))));
+                map.put("alwaysUsers",   new ArrayList<>(Arrays.asList((alwaysUsers   == null ? "" : alwaysUsers)  .split(";"))));
+                map.put("alwaysRoles",   new ArrayList<>(Arrays.asList((alwaysRoles   == null ? "" : alwaysRoles)  .split(";"))));
                 settings.put(id, map);
             }
             rs.close();
@@ -72,34 +70,45 @@ public class Main extends CommandHost {
         }
     }
 
-    private void voiceJoin(Channel c, Guild g, Member m) {
-        if (c.getMembers().size() < 2) return;
+    private void voiceJoin(VoiceChannel c, Guild g, Member m) {
+        if (c.getMembers().size() < 2 || !g.getSelfMember().hasPermission(Permission.MANAGE_CHANNEL)) return;
+
+        if (g.getTextChannelsByName(c.getName().replace(" ", "-") + "-voice", true).size() != 0) return;
 
         String name = c.getName().replace(" ", "-") + "-voice";
+        long gid = g.getIdLong();
 
         if (g.getTextChannelsByName(name, true).size() == 0) {
-            Channel ch = g.getController().createTextChannel(name).complete();
-            ch.getManager().setParent(c.getParent()).queue();
+            ChannelAction ch = g.getController().createTextChannel(name).setParent(c.getParent())
+                    .addPermissionOverride(g.getPublicRole(), 0, Permission.MESSAGE_READ.getRawValue());
 
-            ch.getMemberPermissionOverrides().forEach(q -> q.getManagerUpdatable().reset());
+            List<String> alwaysUsers = settings.get(gid).get("alwaysUsers");
+            if (alwaysUsers.size() != 1 || !alwaysUsers.get(0).equals("")) for (String id : alwaysUsers)
+                ch = ch.addPermissionOverride(g.getMemberById(id), Permission.MESSAGE_READ.getRawValue(), 0);
 
+            List<String> alwaysRoles = settings.get(gid).get("alwaysRoles");
+            if (alwaysRoles.size() != 1 || !alwaysRoles.get(0).equals("")) for (String id : alwaysRoles)
+                ch = ch.addPermissionOverride(g.getRoleById(id), Permission.MESSAGE_READ.getRawValue(), 0);
 
-            ch.createPermissionOverride(g.getRoleById(roles.getString("member"))).setDeny (Permission.MESSAGE_READ).queue();
-            ch.createPermissionOverride(g.getRoleById(roles.getString("admin"))) .setAllow(Permission.MESSAGE_READ).queue();
-            ch.createPermissionOverride(g.getRoleById(roles.getString("bot")))   .setAllow(Permission.MESSAGE_READ).queue();
+            for (Member mbr : c.getMembers())
+                ch = ch.addPermissionOverride(mbr, Permission.MESSAGE_READ.getRawValue(), 0);
 
-            c.getMembers().forEach(q -> ch.createPermissionOverride(q).setAllow(Permission.MESSAGE_READ).queue());
+            ch.queue();
         } else {
+            if (settings.get(gid).get("disallowUsers").contains(m.getUser().getId())) return;
+            for (Role r : m.getRoles())
+                if (settings.get(gid).get("disallowRoles").contains(r.getId())) return;
             g.getTextChannelsByName(name, true).get(0).createPermissionOverride(m).setAllow(Permission.MESSAGE_READ).queue();
         }
     }
 
     private void voiceLeave(VoiceChannel ch, Guild g, Member m) {
+        if (!g.getSelfMember().hasPermission(Permission.MANAGE_CHANNEL)) return;
         String name = ch.getName().replace(" ", "-") + "-voice";
         List<TextChannel> c = g.getTextChannelsByName(name, true);
         if (c.size() == 1) {
             if (ch.getMembers().size() < 2) c.get(0).delete().queue();
-            else c.get(0).getPermissionOverride(m).getManagerUpdatable().getPermissionOverride().delete().queue();
+            else c.get(0).getPermissionOverride(m).getManager().getPermissionOverride().delete().queue();
         }
     }
 
@@ -120,29 +129,32 @@ public class Main extends CommandHost {
         voiceJoin(event.getChannelJoined(), event.getGuild(), event.getMember());
     }
 
-    //
-    // TODO: move V
-    //
+    @Override
+    public void onEnabled(long gid, TextChannel c) {
+        if (!modulebot.voiceText.Main.settings.containsKey(gid)) {
+            try {
+                PreparedStatement st = modulebot.main.Main.conn.prepareStatement("INSERT INTO voiceText VALUES (?, '', '', '', '')");
+                st.setLong(1, gid);
+                st.execute();
+                st.close();
 
-    private static JSONObject channels;
-    private static JSONObject config;
-    @Override
-    public void onGuildMemberJoin(GuildMemberJoinEvent event) {
-        if (event.getMember().getUser().isBot()) {
-            event.getGuild().getController().addRolesToMember(event.getMember(), event.getGuild().getRoleById(roles.getString("bot"))).queue();
-            event.getGuild().getTextChannelById(channels.getString("default")).sendMessage("A bot (" + event.getMember().getEffectiveName() + ") was added to the guild").queue();
-        } else {
-            event.getGuild().getController().addRolesToMember(event.getMember(), event.getGuild().getRoleById(roles.getString("member"))).queue();
-            event.getGuild().getTextChannelById(channels.getString("default")).sendMessage("" +
-                    "Welcome, " + event.getMember().getUser().getAsMention() + "! Please read <#" + channels.getString("readme") + ">!\n" +
-                    "To get a language role just do `" + config.getString("role cmd") + "`\n" +
-                    "To see a list of available roles, do `" + config.getString("roles cmd") + "`!"
-            ).queue();
-            event.getGuild().getTextChannelById(channels.getString("admin")).sendMessage(event.getMember().getEffectiveName() + " joined the guild!").queue();
+                if (c != null) {
+                    c.sendMessage("You should configure guild's voice-text settings using `" + modulebot.main.Main.prefix.get(gid) + "voice-text config {args}`").queue();
+                }
+            } catch (SQLException e) {
+                if (c != null) {
+                    c.sendMessage("Error setting up database, please contact bot owner: \n" +
+                            c.getJDA().asBot().getApplicationInfo().complete().getOwner().getAsMention() +
+                            "**ERROR**: command threw a " + e.getClass().getSimpleName()).queue();
+                }
+                e.printStackTrace();
+            }
         }
-    }
-    @Override
-    public void onGuildMemberLeave(GuildMemberLeaveEvent event) {
-        event.getGuild().getTextChannelById(channels.getString("admin")).sendMessage(event.getMember().getEffectiveName() + " left the guild!").queue();
+
+        if (c != null) {
+            if (!c.getGuild().getSelfMember().hasPermission(Permission.MANAGE_CHANNEL)) {
+                c.sendMessage("The bot needs the Manage Channels permission to work correctly").queue();
+            }
+        }
     }
 }
